refactor: Rename package code_check -> codecheck
