+++ app/src/androidTest/kotlin/jp/co/yumemi/android/codecheck/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.codecheck
 
 import androidx.test.platform.app.InstrumentationRegistry
 import androidx.test.ext.junit.runners.AndroidJUnit4
