+++ app/src/main/kotlin/jp/co/yumemi/android/codecheck/topActivity.kt
@@ -1,7 +1,7 @@
 /*
  * Copyright © 2021 YUMEMI Inc. All rights reserved.
  */
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.codecheck
 
 import androidx.appcompat.app.AppCompatActivity
 import java.util.*
