+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/oneFragment"
-        android:name="jp.co.yumemi.android.code_check.OneFragment"
+        android:name="jp.co.yumemi.android.codecheck.OneFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_one">
         <action
@@ -17,12 +17,12 @@
 
     <fragment
         android:id="@+id/twoFragment"
-        android:name="jp.co.yumemi.android.code_check.TwoFragment"
+        android:name="jp.co.yumemi.android.codecheck.TwoFragment"
         android:label="@string/app_name"
         tools:layout="@layout/fragment_two">
         <argument
             android:name="item"
-            app:argType="jp.co.yumemi.android.code_check.item" />
+            app:argType="jp.co.yumemi.android.codecheck.item" />
     </fragment>
 
 </navigation>
