+++ app/src/test/kotlin/jp/co/yumemi/android/codecheck/ExampleUnitTest.kt
@@ -1,4 +1,4 @@
-package jp.co.yumemi.android.code_check
+package jp.co.yumemi.android.codecheck
 
 import org.junit.Test
